# openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "0.0.1-11-PAGOPA-1647-pm-ingestion"
  },
  "servers": [
    {
      "url": "http://localhost",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/extraction/data": {
      "post": {
        "tags": [
          "PM extraction data REST APIs"
        ],
        "summary": "Request for data extraction from the PM.",
        "operationId": "pmDataExtraction",
        "parameters": [
          {
            "name": "pmExtractionType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "CARD",
              "enum": [
                "CARD",
                "BPAY",
                "PAYPAL"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataExtractionOptionsModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "409": {
            "description": "Conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "401": {
            "description": "Wrong or missing function key."
          },
          "200": {
            "description": "Request paid."
          },
          "400": {
            "description": "Malformed request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          },
          "500": {
            "description": "Service unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "DataExtractionOptionsModel": {
        "type": "object",
        "properties": {
          "taxCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "creationDateFrom": {
            "type": "string",
            "description": "if provided use the format yyyy-MM-dd"
          },
          "creationDateTo": {
            "type": "string",
            "description": "if provided use the format yyyy-MM-dd"
          }
        }
      },
      "ProblemJson": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "A short, summary of the problem type. Written in english and readable for engineers (usually not suited for non technical stakeholders and not localized); example: Service Unavailable"
          },
          "status": {
            "maximum": 600,
            "minimum": 100,
            "type": "integer",
            "description": "The HTTP status code generated by the origin server for this occurrence of the problem.",
            "format": "int32",
            "example": 200
          },
          "detail": {
            "type": "string",
            "description": "A human readable explanation specific to this occurrence of the problem.",
            "example": "There was an error processing the request"
          }
        }
      }
    }
  }
}